Remove cli provision file and redundant comments
